% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratLowDim.R
\name{SeuratLowDim}
\alias{SeuratLowDim}
\title{Create a Low dimensional Seurat object from scaled seurat object}
\usage{
SeuratLowDim(scaled_seurat_object, ...)
}
\arguments{
\item{scaled_seurat_object}{A scaled Seurat object.}

\item{...}{Additional arguments to be passed to convert to low dimensional data frame.}
}
\value{
A Seurat object.
}
\description{
This function converts the transformed data to low-dimensional data for downstream analysis.
}
