% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratMarkers.R
\name{SeuratMarkers}
\alias{SeuratMarkers}
\title{A thresholded markers list for better calculation of DE genes}
\usage{
SeuratMarkers(lowdim_seurat_object)
}
\arguments{
\item{lowdim_seurat_object}{Seurat object with cluster information}
}
\value{
A list
}
\description{
A thresholded markers list for better calculation of DE genes
}
