% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeuratPreprocess.R
\name{SeuratPreprocess}
\alias{SeuratPreprocess}
\title{Create a Seurat object from count data}
\usage{
SeuratPreprocess(counts_data, ...)
}
\arguments{
\item{counts_data}{A matrix or data frame of count data.}

\item{...}{Additional arguments to be passed to Seurat::CreateSeuratObject.}
}
\value{
A Seurat object.
}
\description{
This function creates a Seurat object from count data using Seurat::CreateSeuratObject.
}
